problema1/requirements.txt
cassandra-driver==3.29.3
click==8.3.0
dnspython==2.8.0
geomet==1.1.0
numpy==2.3.4
pandas==2.3.3
psycopg2-binary==2.9.11
pymongo==4.15.4
python-dateutil==2.9.0.post0
pytz==2025.2
redis==7.0.1
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
